move json tests to subfolder
